refactor-changed pages name to end with Page
