+++ weather-app-frontend/src/main.tsx
@@ -2,10 +2,16 @@ import React from "react";
 import ReactDOM from "react-dom/client";
 import App from "./App.tsx";
 import { ChakraProvider, ColorModeScript } from "@chakra-ui/react";
-// import { ColorModeScript } from "@chakra-ui/color-mode";
 import { Provider } from "react-redux";
 import store from "../src/features/store";
 import { chakraCustomTheme } from "./layout/theme.ts";
+// import { createBrowserRouter } from "react-router-dom";
+
+// const router = createBrowserRouter([
+//   {
+//     path: "/noam-rivlin-19-11-23/",
+//   },
+// ]);
 
 ReactDOM.createRoot(document.getElementById("root")!).render(
   <React.StrictMode>
