+++ weather-app-frontend/src/pages/FavoritesPage.tsx
